Separate wrapper out from styled menu
